GameObject hierarchy and transform dirty flag
Fixed Dictionary memory leak.
Added function to create game object.
Added game object parenting.
Added a bit of logic to ResourceManager to check for the Data folder.
Var Wrapper with some kind of get set functionality??
Now Copies the data folder and it's contents to output directory... why was this not a thing to begin with?